% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{add_edge}
\alias{add_edge}
\title{add_edge}
\usage{
add_edge(object, vertex_pairs)
}
\arguments{
\item{object, }{Object of class Network.}

\item{vertex_pairs, }{numeric matrix that store the pairs of vertexs}
}
\value{
a new object that store the network which added edges
}
\description{
This function is to Create a method to add edges
}
\author{
Tong Yin
}
