% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMI_met_cis_network.R
\name{CMI_met_cis_network}
\alias{CMI_met_cis_network}
\title{Conditional mutual information learning the methylation cis-acting regulation network}
\usage{
CMI_met_cis_network(met_data_matrix,exp_data_matrix,gene_list,distance=300000,
ref_gene_bed,ref_CpGs_bed,outfiledir=NA,pvalue_cut=0.001,core_num=1,permutation_times=100)
}
\arguments{
\item{met_data_matrix}{a numeric matrix containing CpGs methylation data where columns contain samples and rows contain variables(probe site)}

\item{exp_data_matrix}{a numeric matrix containing gene expression data where columns contain samples and rows contain variables(gene site)}

\item{gene_list}{a vector containing the names of target genes}

\item{distance}{integer specifying the upstream/downstream genome range to be analyzed}

\item{ref_gene_bed}{a data.frame containing reference gene coorinate with five columns named "name", "chr", "start", "end" and "strand". The coordinates of genes in exp_data_matrix are required to be included in this data.frame.}

\item{ref_CpGs_bed}{a data.frame containing reference CpGS coorinate with four columns names "name", "chr", "start" and "end". The coordinates of CpGs/probes in met_data_matrix are required to be included in this data.frame.}

\item{outfiledir}{a string of file directory to store the result files. If the parameter is not specified, the log file directory will be get by \code{getwd()}.}

\item{pvalue_cut}{the cutoff of pvalue. The default is 0.01.}

\item{core_num}{the cpu number using for parallel computation in PC_para}

\item{permutation_times}{the number of times of permutation to calculate the pvalue}
}
\value{
the adjacency matrix of the network with value of 0 and 1. 1 means that there is an edge between the rowname and colname of the element.
And 0 means there is no edge.
}
\description{
This function is to infer the cis-acting regulatory network between DNA methylation and gene expression
}
\examples{
data("TCGA_STAD_data")
gene_name<-"MLH1"
\dontrun{
network<-CMI_met_cis_network(met_data_matrix=STAD_met_data_matrix,
exp_data_matrix=STAD_exp_data_matrix,gene_list=gene_name,distance=300000,
ref_gene_bed=STAD_ref_gene_bed,ref_CpGs_bed=STAD_ref_CpGs_bed,pvalue_cut=0.00001,
permutation_times=20)
}
}
\author{
Tong Yin
}
