% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{delete_edge}
\alias{delete_edge}
\title{delete_edge}
\usage{
delete_edge(object, vertex_pairs)
}
\arguments{
\item{object, }{Object of class Network.}

\item{vertex_pairs, }{numeric matrix that store the pairs of vertexs}
}
\value{
a new object that store the network which deleted edges
}
\description{
This function is to Create a method to delete edges
}
\author{
Tong Yin
}
