% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMI_met_cis_statistic.R
\name{generate_regulator_info}
\alias{generate_regulator_info}
\title{generate_regulator_info}
\usage{
generate_regulator_info(met_data_matrix,exp_data_matrix,gene_list,
outfiledir=NA,ref_gene_bed,ref_CpGs_bed)
}
\arguments{
\item{met_data_matrix}{a SummarizedExperiment object or a numeric matrix containing CpGs methylation data where columns contain samples and rows contain variables(probe site)}

\item{exp_data_matrix}{a SummarizedExperiment object or a numeric matrix containing gene expression data where columns contain samples and rows contain variables(gene site)}

\item{gene_list}{a vector containing the names of target genes}

\item{outfiledir}{a string of file directory to store the result files. If the parameter is not specified, the log file directory will be get by \code{getwd()}.}

\item{ref_gene_bed}{a GRanges object or a data.frame containing reference gene coorinate with five columns named "name", "chr", "start", "end" and "strand". The coordinates of genes in exp_data_matrix are required to be included in this data.frame.}

\item{ref_CpGs_bed}{a GRanges object or a data.frame containing reference CpGS coorinate with four columns names "name", "chr", "start" and "end". The coordinates of CpGs/probes in met_data_matrix are required to be included in this data.frame.}

\item{promoter_upstrm}{a distance number, define the promoter region that in upstream to TSS}

\item{promoter_dwstrm}{a distance number, define the promoter region that in downstream to TSS}
}
\value{
data.frame containing information of direct and indirect regulators
}
\description{
This function is to integrate the regulator information in the gene_regulator_info.txt file
}
\examples{

data("TCGA_STAD_data")
gene_name<-"MLH1"
\dontrun{
generate_regulator_info(met_data_matrix=STAD_met_data_matrix,
exp_data_matrix=STAD_exp_data_matrix,gene_list=gene_name,
ref_gene_bed=STAD_ref_gene_bed,ref_CpGs_bed=STAD_ref_CpGs_bed)
}
}
\seealso{
the usage of CMI_met_cis_network
}
\author{
Tong Yin
}
