% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{get_edge_number}
\alias{get_edge_number}
\title{get_edge_number}
\usage{
get_edge_number(object)
}
\arguments{
\item{object, }{Object of class Network.}
}
\value{
number
}
\description{
This function is to get number of edges
}
\author{
Tong Yin
}
