% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{get_partners}
\alias{get_partners}
\title{get_partners}
\usage{
get_partners(object, vertex)
}
\arguments{
\item{object, }{Object of class Network.}

\item{vertex, }{a vector that store the names of nodes}
}
\value{
a vector of partners
}
\description{
This function is to Create a method to get partners for any vertex
}
\author{
Tong Yin
}
