% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{get_sharing_partners}
\alias{get_sharing_partners}
\title{get_sharing_partners}
\usage{
get_sharing_partners(object, vertex1, vertex2)
}
\arguments{
\item{object, }{Object of class Network.}

\item{vertex1, }{vector that store the name of the first node}

\item{vertex2, }{vector that store the name of the secode node}
}
\value{
a vector of partners
}
\description{
This function is to Create a method to get sharing partners for vertex
}
\author{
Tong Yin
}
