% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{get_vertex_number}
\alias{get_vertex_number}
\title{get_vertex_number}
\usage{
get_vertex_number(object)
}
\arguments{
\item{object, }{Object of class Network.}
}
\value{
number
}
\description{
This function is to get number of vertexs
}
\author{
Tong Yin
}
