% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{HNSC_control_id}
\alias{HNSC_control_id}
\title{the sample ids of normal samples in TCGA HNSC dataset}
\format{A vector of character}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_HNSC_data)
}
\value{
A vector of sample ids
}
\description{
A vector containing sample ids
}
\references{
Cancer Genome Atlas Network. "Comprehensive genomic characterization of head and neck squamous cell carcinomas." Nature 517.7536 (2015): 576-582.
}
\keyword{datasets}
