% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{HNSC_exp_data_matrix}
\alias{HNSC_exp_data_matrix}
\title{expression values of genes in HNSC samples}
\format{A matrix of numeric values}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_HNSC_data)
}
\value{
The log2 expression matrix for head and neck cancer samples
}
\description{
A data matrix containing the log2 transformed expression levels of genes in
head and neck cancer samples from TCGA
}
\references{
Cancer Genome Atlas Network. "Comprehensive genomic characterization of head and neck squamous cell carcinomas." Nature 517.7536 (2015): 576-582.
}
\keyword{datasets}
