% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{HNSC_met_data_matrix}
\alias{HNSC_met_data_matrix}
\title{methylation values of CpGs in HNSC samples}
\format{A matrix of numeric values}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_HNSC_data)
}
\value{
The methylation beta matrix for head and neck cancer samples
}
\description{
A data matrix containing the methylation beta values of CpGs in head and neck cancer
samples from TCGA
}
\references{
Cancer Genome Atlas Network. "Comprehensive genomic characterization of head and neck squamous cell carcinomas." Nature 517.7536 (2015): 576-582.
}
\keyword{datasets}
