% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{HNSC_ref_CpGs_bed}
\alias{HNSC_ref_CpGs_bed}
\title{the CpGs coordinates in bed format}
\format{A data.frame}
\source{
\url{https://genome.ucsc.edu}
}
\usage{
data(TCGA_HNSC_data)
}
\value{
A data.frame containing genome coordinates of CpGs
\describe{
  \item{name}{name of CpGs}
  \item{chr}{the chromosome id like chr1, chr2, chr3 ...}
  \item{start}{the starting coordinate of CpGs}
  \item{end}{the ending coordinate of CpGs}
}
}
\description{
A data.frame containing the chromosome coordinate information of all the CpGs
nearby the target genes
}
\keyword{datasets}
