% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{HNSC_sample_class}
\alias{HNSC_sample_class}
\title{the classes of samples}
\format{A data.frame}
\source{
\url{https://genome.ucsc.edu}
}
\usage{
data(TCGA_HNSC_data)
}
\value{
sample classification information in two column data.frame
\describe{
  \item{sample_id}{name of samples, should be exactly the same as the colnames of data matrix}
  \item{class}{control, tumor, stages, or tumor subtypes}
}
}
\description{
A data.frame containing the classes(control,tumor,stages or subtypes) information of samples
}
\keyword{datasets}
