% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMI_met_cis_statistic.R
\name{merge_regulator_info}
\alias{merge_regulator_info}
\title{merge_regulator_info}
\usage{
merge_regulator_info(gene_list,outfiledir=NA,statisticfiledir=NA,ref_gene_bed)
}
\arguments{
\item{gene_list}{a vector containing the names of target genes}

\item{outfiledir}{a string of file directory to store the result files. If the parameter is not specified, the log file directory will be get by \code{getwd()}.}

\item{statisticfiledir}{summary directory to store merged result. If the parameter is not specified, the file directory will be get by \code{getwd()}.}

\item{ref_gene_bed}{a data.frame containing reference gene coorinate with five columns named "name", "chr", "start", "end" and "strand". The coordinates of genes in exp_data_matrix are required to be included in this data.frame.}
}
\value{
numbers of direct and indirect regulators
}
\description{
This function is to merge regulation information for multiple genes
}
\examples{
\dontrun{
data("TCGA_LUAD_data")
gene_list<-rownames(LUAD_exp_data_matrix)[41:50]

network<-CMI_met_cis_network(met_data_matrix=LUAD_met_data_matrix,
exp_data_matrix=LUAD_exp_data_matrix,gene_list=gene_list,distance=300000,
ref_gene_bed=LUAD_ref_gene_bed,ref_CpGs_bed=LUAD_ref_CpGs_bed,
core_num=1,permutation_times=20)

generate_regulator_info(met_data_matrix=LUAD_met_data_matrix,
exp_data_matrix=LUAD_exp_data_matrix,gene_list=gene_list,
ref_gene_bed=LUAD_ref_gene_bed,ref_CpGs_bed=LUAD_ref_CpGs_bed)

merge_regulator_info(gene_list=gene_list,ref_gene_bed=LUAD_ref_gene_bed)
}
}
\author{
Tong Yin
}
