% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{class}
\name{Network}
\alias{Network}
\title{Network
An S4 class to store the network in adjacent matrix}
\description{
Network
An S4 class to store the network in adjacent matrix
}
\section{Slots}{

\describe{
\item{\code{vertex,}}{a string vector to store the node names}

\item{\code{edges,}}{a numeric matrix with two columns to store the edges}

\item{\code{adj_matrix,}}{a adjacent matrix to store the unidirectional network}

\item{\code{bi_adj_matrix,}}{a adjacent matrix to sotre the bidirectional network}
}}

\author{
Tong Yin
}
