% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{read_adj_matrix}
\alias{read_adj_matrix}
\title{read_adj_matrix}
\usage{
read_adj_matrix(object, adj_matrix)
}
\arguments{
\item{object, }{Object of class Network.}

\item{adj_matrix, }{an adjacent matrix to load}
}
\value{
the new object loaded with adjacent matrix
}
\description{
This function is to Create a method to read genes annotation data
}
\author{
Tong Yin
}
