% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{STAD_exp_data_matrix}
\alias{STAD_exp_data_matrix}
\title{expression values of genes in STAD samples}
\format{A matrix of numeric values}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_STAD_data)
}
\value{
The log2 expression matrix for gastric cancer samples
}
\description{
A data matrix containing the log2 transformed expression levels of genes in
gastric cancer samples from TCGA
}
\references{
Cancer Genome Atlas Research Network. "Comprehensive molecular characterization of gastric adenocarcinoma." Nature 513.7517 (2014): 202-209.
}
\keyword{datasets}
