% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{STAD_met_data_matrix}
\alias{STAD_met_data_matrix}
\title{methylation values of CpGs in STAD samples}
\format{A matrix of numeric values}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_STAD_data)
}
\value{
The methylation beta matrix for gastric cancer samples
}
\description{
A data matrix containing the methylation beta values of CpGs in gastric cancer
samples from TCGA
}
\references{
Cancer Genome Atlas Research Network. "Comprehensive molecular characterization of gastric adenocarcinoma." Nature 513.7517 (2014): 202-209.
}
\keyword{datasets}
