% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{STAD_ref_gene_bed}
\alias{STAD_ref_gene_bed}
\title{the gene coordinates in bed format}
\format{A data.frame}
\source{
\url{https://genome.ucsc.edu}
}
\usage{
data(TCGA_STAD_data)
}
\value{
A data.frame containing genome coordinates of genes
\describe{
  \item{name}{name of genes}
  \item{chr}{the chromosome id like chr1, chr2, chr3 ...}
  \item{start}{the starting coordinate of genes}
  \item{end}{the ending coordinate of genes}
  \item{strand}{the strand of genes}
}
}
\description{
A data.frame containing the chromosome coordinate information of all the genes
nearby the target genes
}
\keyword{datasets}
