% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{STAD_tumor_id}
\alias{STAD_tumor_id}
\title{the sample ids of tumor samples in TCGA STAD dataset}
\format{A vector of character}
\source{
\url{http://cancergenome.nih.gov/}
}
\usage{
data(TCGA_STAD_data)
}
\value{
A vector of sample ids
}
\description{
A vector containing sample ids
}
\references{
Cancer Genome Atlas Research Network. "Comprehensive molecular characterization of gastric adenocarcinoma." Nature 513.7517 (2014): 202-209.
}
\keyword{datasets}
