% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data_2.R
\docType{data}
\name{TCGA_HNSC_data}
\alias{TCGA_HNSC_data}
\alias{HNSC_met_data_matrix}
\alias{HNSC_exp_data_matrix}
\alias{HNSC_control_id}
\alias{HNSC_tumor_id}
\alias{HNSC_ref_gene_bed}
\alias{HNSC_ref_CpGs_bed}
\alias{HNSC_sample_class}
\title{methylation and expression data from TCGA HNSC dataset}
\format{matrix, data.frame}
\source{
\url{http://cancergenome.nih.gov/}

\url{https://genome.ucsc.edu}
}
\usage{
data(TCGA_HNSC_data)

HNSC_exp_data_matrix

HNSC_control_id

HNSC_tumor_id

HNSC_ref_gene_bed

HNSC_ref_CpGs_bed

HNSC_sample_class
}
\value{
The methylation beta matrix, and expression log2 matrix
}
\description{
methylation and expression data from TCGA HNSC dataset
}
\details{
HNSC_met_data_matrix: methylation values of CpGs in HNSC samples.
A data matrix containing the methylation beta values of CpGs in gastric cancer samples from TCGA

HNSC_exp_data_matrix: expression values of genes in HNSC samples.
A data matrix containing the log2 transformed expression levels of genes in
gastric cancer samples from TCGA

HNSC_control_id: the sample ids of normal samples in TCGA HNSC dataset.
A vector containing sample ids

HNSC_tumor_id: the sample ids of tumor samples in TCGA HNSC dataset.
A vector containing sample ids

HNSC_ref_gene_bed: the gene coordinates in bed format.
A data.frame containing the chromosome coordinate information of all the genes
nearby the target genes

HNSC_ref_CpGs_bed: the CpGs coordinates in bed format.
A data.frame containing the chromosome coordinate information of all the CpGs
nearby the target genes

HNSC_sample_class:
a data.frame containing the classes(control,tumor,stages or subtypes) information of samples
}
\references{
Cancer Genome Atlas Research Network. "Comprehensive molecular characterization of gastric adenocarcinoma." Nature 513.7517 (2014): 202-209.
}
\keyword{datasets}
