% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{TCGA_STAD_data}
\alias{TCGA_STAD_data}
\alias{STAD_met_data_matrix}
\alias{STAD_exp_data_matrix}
\alias{STAD_control_id}
\alias{STAD_tumor_id}
\alias{STAD_ref_gene_bed}
\alias{STAD_ref_CpGs_bed}
\alias{STAD_sample_class}
\title{methylation and expression data from TCGA STAD dataset}
\format{matrix, data.frame}
\source{
\url{http://cancergenome.nih.gov/}

\url{https://genome.ucsc.edu}
}
\usage{
data(TCGA_STAD_data)

STAD_exp_data_matrix

STAD_control_id

STAD_tumor_id

STAD_ref_gene_bed

STAD_ref_CpGs_bed

STAD_sample_class
}
\value{
The methylation beta matrix, and expression log2 matrix
}
\description{
methylation and expression data from TCGA STAD dataset
}
\details{
STAD_met_data_matrix: methylation values of CpGs in STAD samples.
A data matrix containing the methylation beta values of CpGs in gastric cancer samples from TCGA

STAD_exp_data_matrix: expression values of genes in STAD samples.
A data matrix containing the log2 transformed expression levels of genes in
gastric cancer samples from TCGA

STAD_control_id: the sample ids of normal samples in TCGA STAD dataset.
A vector containing sample ids

STAD_tumor_id: the sample ids of tumor samples in TCGA STAD dataset.
A vector containing sample ids

STAD_ref_gene_bed: the gene coordinates in bed format.
A data.frame containing the chromosome coordinate information of all the genes
nearby the target genes

STAD_ref_CpGs_bed: the CpGs coordinates in bed format.
A data.frame containing the chromosome coordinate information of all the CpGs
nearby the target genes

STAD_sample_class:
a data.frame containing the classes(control,tumor,stages or subtypes) information of samples
}
\references{
Cancer Genome Atlas Research Network. "Comprehensive molecular characterization of gastric adenocarcinoma." Nature 513.7517 (2014): 202-209.
}
\keyword{datasets}
