% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{test_vertex_pairs}
\alias{test_vertex_pairs}
\title{test_vertex_pairs}
\usage{
test_vertex_pairs(object, vertex_pairs)
}
\arguments{
\item{object, }{Object of class Network.}

\item{vertex_pairs, }{numeric matrix that store the pairs of vertexs}
}
\value{
testing information
}
\description{
This function is to test whether the nodes in the input edges are in the network or not
}
\author{
Tong Yin
}
