% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{methods}
\name{update_edge}
\alias{update_edge}
\title{update_edge}
\usage{
update_edge(object)
}
\arguments{
\item{object, }{Object of class Network.}
}
\value{
a new object that store the network which added edges
}
\description{
This function is to Create a method to update edge data from object
}
\author{
Tong Yin
}
