% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Class.R
\docType{class}
\name{Network}
\alias{Network}
\alias{read_adj_matrix}
\alias{get_partners}
\alias{get_sharing_partners}
\alias{delete_edge}
\alias{add_edge}
\alias{update_edge}
\alias{test_vertex_pairs}
\alias{get_vertex_number}
\alias{get_edge_number}
\title{Network
An S4 class to store the network in adjacent matrix}
\usage{
read_adj_matrix(object, adj_matrix)

get_partners(object, vertex)

get_sharing_partners(object, vertex1, vertex2)

delete_edge(object, vertex_pairs)

add_edge(object, vertex_pairs)

update_edge(object)

test_vertex_pairs(object, vertex_pairs)

get_vertex_number(object)

get_edge_number(object)
}
\arguments{
\item{object, }{Object of class Network.}

\item{adj_matrix, }{an adjacent matrix to load}

\item{vertex, }{a vector that store the names of nodes}

\item{vertex1, }{vector that store the name of the first node}

\item{vertex2, }{vector that store the name of the secode node}

\item{vertex_pairs, }{numeric matrix that store the pairs of vertexs}

\item{object, }{Object of class Network.}

\item{object, }{Object of class Network.}

\item{object, }{Object of class Network.}

\item{object, }{Object of class Network.}

\item{vertex_pairs, }{numeric matrix that store the pairs of vertexs}

\item{object, }{Object of class Network.}

\item{object, }{Object of class Network.}

\item{vertex_pairs, }{numeric matrix that store the pairs of vertexs}

\item{object, }{Object of class Network.}

\item{object, }{Object of class Network.}
}
\value{
the new object loaded with adjacent matrix

a vector of partners

a vector of partners

a new object that store the network which deleted edges

a new object that store the network which added edges

a new object that store the network which added edges

testing information

number

number
}
\description{
Network
An S4 class to store the network in adjacent matrix

This function is to Create a method to read genes annotation data

This function is to Create a method to get partners for any vertex

This function is to Create a method to get sharing partners for vertex

This function is to Create a method to delete edges

This function is to Create a method to add edges

This function is to Create a method to update edge data from object

This function is to test whether the nodes in the input edges are in the network or not

This function is to get number of vertexs

This function is to get number of edges
}
\section{Slots}{

\describe{
\item{\code{vertex,}}{a string vector to store the node names}

\item{\code{edges,}}{a numeric matrix with two columns to store the edges}

\item{\code{adj_matrix,}}{a adjacent matrix to store the unidirectional network}

\item{\code{bi_adj_matrix,}}{a adjacent matrix to sotre the bidirectional network}
}}

\author{
Tong Yin

Tong Yin

Tong Yin

Tong Yin

Tong Yin

Tong Yin

Tong Yin

Tong Yin

Tong Yin

Tong Yin
}
\keyword{internal}
